+++ client/pages/index/drafts/list.vue
@@ -44,8 +44,17 @@ export default class DraftListPage extends Vue {
     handleEdit(idx, id) {
         this.$router.push(MyPagePath.draftPages.getEditPath(id));
     }
-    handleDelete(idx, id) {
-        MyDraftAPI.deleteAPI(this.$axios, id);
+    handleDelete(index,id) {
+        this.$confirm("此操作将永久删除该草稿, 是否继续?", "提示", {
+            confirmButtonText: "确定",
+            cancelButtonText: "取消",
+            type: "warning"
+        })
+            .then(async () => {
+                let res = await MyDraftAPI.deleteAPI(this.$axios, id);
+                res.success && this.getData(1)
+            })
+            .catch(() => {});
     }
     created() {
         Bus.$on(`pageChange_${this.pageName}`, this.getData);
@@ -54,14 +63,17 @@ export default class DraftListPage extends Vue {
         this.getData(1);
     }
     getData(page: number) {
-        MyDraftAPI.getTotalNumberAPI(this.$axios).then(res => {
-            if (res.success) {
-                this.totalData = res.data;
-            }
-        });
+        this.getTotalNumber()
         MyDraftAPI.findAllAPI(this.$axios, this.pageSize, page).then(res => {
             if (res.success) {
                 this.tableData = res.data;
+            }
+        });
+    }
+    getTotalNumber(){
+        MyDraftAPI.getTotalNumberAPI(this.$axios).then(res => {
+            if (res.success) {
+                this.totalData = res.data;
             }
         });
     }
